+++ lib/src/fields/form_builder_dropdown.dart
@@ -3,25 +3,209 @@ import 'package:flutter/widgets.dart';
 import 'package:flutter_form_builder/flutter_form_builder.dart';
 
 class FormBuilderDropdown<T> extends FormBuilderField<T> {
+  /// The list of items the user can select.
+  ///
+  /// If the [onChanged] callback is null or the list of items is null
+  /// then the dropdown button will be disabled, i.e. its arrow will be
+  /// displayed in grey and it will not respond to input. A disabled button
+  /// will display the [disabledHint] widget if it is non-null. If
+  /// [disabledHint] is also null but [hint] is non-null, [hint] will instead
+  /// be displayed.
+  final List<DropdownMenuItem<T>> items;
+
+  /// A placeholder widget that is displayed by the dropdown button.
+  ///
+  /// If [value] is null, this widget is displayed as a placeholder for
+  /// the dropdown button's value. This widget is also displayed if the button
+  /// is disabled ([items] or [onChanged] is null) and [disabledHint] is null.
   final Widget hint;
-  final List<DropdownMenuItem> items;
-  final bool isExpanded;
-  final TextStyle style;
-  final bool isDense;
-  final int elevation;
+
+  /// A message to show when the dropdown is disabled.
+  ///
+  /// Displayed if [items] or [onChanged] is null. If [hint] is non-null and
+  /// [disabledHint] is null, the [hint] widget will be displayed instead.
   final Widget disabledHint;
-  final double iconSize;
+
+  /// Called when the dropdown button is tapped.
+  ///
+  /// This is distinct from [onChanged], which is called when the user
+  /// selects an item from the dropdown.
+  ///
+  /// The callback will not be invoked if the dropdown button is disabled.
+  final VoidCallback onTap;
+
+  /// A builder to customize the dropdown buttons corresponding to the
+  /// [DropdownMenuItem]s in [items].
+  ///
+  /// When a [DropdownMenuItem] is selected, the widget that will be displayed
+  /// from the list corresponds to the [DropdownMenuItem] of the same index
+  /// in [items].
+  ///
+  /// {@tool dartpad --template=stateful_widget_scaffold}
+  ///
+  /// This sample shows a `DropdownButton` with a button with [Text] that
+  /// corresponds to but is unique from [DropdownMenuItem].
+  ///
+  /// ```dart
+  /// final List<String> items = <String>['1','2','3'];
+  /// String selectedItem = '1';
+  ///
+  /// @override
+  /// Widget build(BuildContext context) {
+  ///   return Padding(
+  ///     padding: const EdgeInsets.symmetric(horizontal: 12.0),
+  ///     child: DropdownButton<String>(
+  ///       value: selectedItem,
+  ///       onChanged: (String string) => setState(() => selectedItem = string),
+  ///       selectedItemBuilder: (BuildContext context) {
+  ///         return items.map<Widget>((String item) {
+  ///           return Text(item);
+  ///         }).toList();
+  ///       },
+  ///       items: items.map((String item) {
+  ///         return DropdownMenuItem<String>(
+  ///           child: Text('Log $item'),
+  ///           value: item,
+  ///         );
+  ///       }).toList(),
+  ///     ),
+  ///   );
+  /// }
+  /// ```
+  /// {@end-tool}
+  ///
+  /// If this callback is null, the [DropdownMenuItem] from [items]
+  /// that matches [value] will be displayed.
+  final DropdownButtonBuilder selectedItemBuilder;
+
+  /// The z-coordinate at which to place the menu when open.
+  ///
+  /// The following elevations have defined shadows: 1, 2, 3, 4, 6, 8, 9, 12,
+  /// 16, and 24. See [kElevationToShadow].
+  ///
+  /// Defaults to 8, the appropriate elevation for dropdown buttons.
+  final int elevation;
+
+  /// The text style to use for text in the dropdown button and the dropdown
+  /// menu that appears when you tap the button.
+  ///
+  /// To use a separate text style for selected item when it's displayed within
+  /// the dropdown button, consider using [selectedItemBuilder].
+  ///
+  /// {@tool dartpad --template=stateful_widget_scaffold}
+  ///
+  /// This sample shows a `DropdownButton` with a dropdown button text style
+  /// that is different than its menu items.
+  ///
+  /// ```dart
+  /// List<String> options = <String>['One', 'Two', 'Free', 'Four'];
+  /// String dropdownValue = 'One';
+  ///
+  /// @override
+  /// Widget build(BuildContext context) {
+  ///   return Container(
+  ///     alignment: Alignment.center,
+  ///     color: Colors.blue,
+  ///     child: DropdownButton<String>(
+  ///       value: dropdownValue,
+  ///       onChanged: (String newValue) {
+  ///         setState(() {
+  ///           dropdownValue = newValue;
+  ///         });
+  ///       },
+  ///       style: TextStyle(color: Colors.blue),
+  ///       selectedItemBuilder: (BuildContext context) {
+  ///         return options.map((String value) {
+  ///           return Text(
+  ///             dropdownValue,
+  ///             style: TextStyle(color: Colors.white),
+  ///           );
+  ///         }).toList();
+  ///       },
+  ///       items: options.map<DropdownMenuItem<String>>((String value) {
+  ///         return DropdownMenuItem<String>(
+  ///           value: value,
+  ///           child: Text(value),
+  ///         );
+  ///       }).toList(),
+  ///     ),
+  ///   );
+  /// }
+  /// ```
+  /// {@end-tool}
+  ///
+  /// Defaults to the [TextTheme.subtitle1] value of the current
+  /// [ThemeData.textTheme] of the current [Theme].
+  final TextStyle style;
+
+  /// The widget to use for the drop-down button's icon.
+  ///
+  /// Defaults to an [Icon] with the [Icons.arrow_drop_down] glyph.
   final Widget icon;
+
+  /// The color of any [Icon] descendant of [icon] if this button is disabled,
+  /// i.e. if [onChanged] is null.
+  ///
+  /// Defaults to [Colors.grey.shade400] when the theme's
+  /// [ThemeData.brightness] is [Brightness.light] and to
+  /// [Colors.white10] when it is [Brightness.dark]
   final Color iconDisabledColor;
+
+  /// The color of any [Icon] descendant of [icon] if this button is enabled,
+  /// i.e. if [onChanged] is defined.
+  ///
+  /// Defaults to [Colors.grey.shade700] when the theme's
+  /// [ThemeData.brightness] is [Brightness.light] and to
+  /// [Colors.white70] when it is [Brightness.dark]
   final Color iconEnabledColor;
+
+  /// The size to use for the drop-down button's down arrow icon button.
+  ///
+  /// Defaults to 24.0.
+  final double iconSize;
+
+  /// Reduce the button's height.
+  ///
+  /// By default this button's height is the same as its menu items' heights.
+  /// If isDense is true, the button's height is reduced by about half. This
+  /// can be useful when the button is embedded in a container that adds
+  /// its own decorations, like [InputDecorator].
+  final bool isDense;
+
+  /// Set the dropdown's inner contents to horizontally fill its parent.
+  ///
+  /// By default this button's inner width is the minimum size of its contents.
+  /// If [isExpanded] is true, the inner width is expanded to fill its
+  /// surrounding container.
+  final bool isExpanded;
+
+  /// If null, then the menu item heights will vary according to each menu item's
+  /// intrinsic height.
+  ///
+  /// The default value is [kMinInteractiveDimension], which is also the minimum
+  /// height for menu items.
+  ///
+  /// If this value is null and there isn't enough vertical room for the menu,
+  /// then the menu's initial scroll offset may not align the selected item with
+  /// the dropdown button. That's because, in this case, the initial scroll
+  /// offset is computed as if all of the menu item heights were
+  /// [kMinInteractiveDimension].
+  final double itemHeight;
+
+  /// The color for the button's [Material] when it has the input focus.
+  final Color focusColor;
+
+  /// {@macro flutter.widgets.Focus.autofocus}
+  final bool autofocus;
+
+  /// The background color of the dropdown.
+  ///
+  /// If it is not provided, the theme's [ThemeData.canvasColor] will be used
+  /// instead.
+  final Color dropdownColor;
+
   final bool allowClear;
   final Widget clearIcon;
-  final VoidCallback onTap;
-  final bool autofocus;
-  final Color dropdownColor;
-  final Color focusColor;
-  final double itemHeight;
-  final List<Widget> Function(BuildContext) selectedItemBuilder;
 
   FormBuilderDropdown({
     Key key,
@@ -35,7 +219,7 @@ class FormBuilderDropdown<T> extends FormBuilderField<T> {
     ValueTransformer valueTransformer,
     bool enabled = true,
     FormFieldSetter onSaved,
-    bool autovalidate = false,
+    AutovalidateMode autovalidateMode = AutovalidateMode.disabled,
     VoidCallback onReset,
     FocusNode focusNode,
     @required this.items,
@@ -65,7 +249,7 @@ class FormBuilderDropdown<T> extends FormBuilderField<T> {
             valueTransformer: valueTransformer,
             onChanged: onChanged,
             readOnly: readOnly,
-            autovalidate: autovalidate,
+            autovalidateMode: autovalidateMode,
             onSaved: onSaved,
             enabled: enabled,
             onReset: onReset,
@@ -81,6 +265,7 @@ class FormBuilderDropdown<T> extends FormBuilderField<T> {
                       ? decoration.floatingLabelBehavior
                       : FloatingLabelBehavior.always,
                 ),
+                isEmpty: state.value == null,
                 child: Row(
                   children: <Widget>[
                     Expanded(