+++ lib/src/widgets/grouped_radio.dart
@@ -1,8 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_form_builder/flutter_form_builder.dart';
 
-import 'grouped_checkbox.dart';
-
 class GroupedRadio<T> extends StatefulWidget {
   /// A list of string that describes each checkbox. Each item must be distinct.
   final List<FormBuilderFieldOption> options;
@@ -17,7 +15,7 @@ class GroupedRadio<T> extends StatefulWidget {
   final List<T> disabled;
 
   /// Specifies the orientation of the elements in itemList.
-  final GroupedRadioOrientation orientation;
+  final OptionsOrientation orientation;
 
   /// Called when the value of the checkbox group changes.
   final ValueChanged<T> onChanged;
@@ -219,18 +217,22 @@ class _GroupedRadioState<T> extends State<GroupedRadio<T>> {
     for (var i = 0; i < widget.options.length; i++) {
       widgetList.add(item(i));
     }
-    if (widget.orientation == GroupedRadioOrientation.vertical) {
+    if (widget.orientation == OptionsOrientation.vertical) {
       for (final item in widgetList) {
         content.add(Row(children: <Widget>[item]));
       }
       finalWidget = SingleChildScrollView(
-          scrollDirection: Axis.vertical, child: Column(children: content));
-    } else if (widget.orientation == GroupedRadioOrientation.horizontal) {
+        scrollDirection: Axis.vertical,
+        child: Column(children: content),
+      );
+    } else if (widget.orientation == OptionsOrientation.horizontal) {
       for (final item in widgetList) {
         content.add(Column(children: <Widget>[item]));
       }
       finalWidget = SingleChildScrollView(
-          scrollDirection: Axis.horizontal, child: Row(children: content));
+        scrollDirection: Axis.horizontal,
+        child: Row(children: content),
+      );
     } else {
       finalWidget = SingleChildScrollView(
         child: Wrap(
@@ -286,11 +288,11 @@ class _GroupedRadioState<T> extends State<GroupedRadio<T>> {
       children: <Widget>[
         if (widget.controlAffinity == ControlAffinity.leading) ...[
           control,
-          label
+          Flexible(child: label),
         ],
         if (widget.controlAffinity == ControlAffinity.trailing) ...[
-          label,
-          control
+          Flexible(child: label),
+          control,
         ],
         if (widget.separator != null &&
             widget.options[index] != widget.options.last)
@@ -299,6 +301,3 @@ class _GroupedRadioState<T> extends State<GroupedRadio<T>> {
     );
   }
 }
-
-enum GroupedRadioOrientation { horizontal, vertical, wrap }
-// enum ControlAffinity { leading, trailing }